created initial unit tests for package.
